更新 `Console` 相关程序说明
