+++ src/console/console/utils.py
@@ -1,7 +1,36 @@
-from typing import List
+import asyncio
+import time
+from threading import Thread
+from typing import List, Optional
 
 from src.const.exceptions import CommandExit
-from src.utils import Style
+from src.utils import Queue, Style
+
+
+class InputQueue(object):
+    _queue: Queue[str]
+    _running: bool
+    _thread: Thread
+
+    def _run(self) -> None:
+        while self._running:
+            try:
+                self._queue.put(input())
+            except EOFError:
+                self._running = False
+            time.sleep(0.05)
+
+    def start(self) -> None:
+        self._queue = Queue(0)
+        self._running = True
+        self._thread = Thread(target=self._run, daemon=True)
+        self._thread.start()
+
+    async def get(self) -> Optional[str]:
+        while self._running:
+            if not self._queue.empty():
+                return self._queue.get()
+            await asyncio.sleep(0.05)
 
 
 def parse_cmd(cmd: str) -> List[str]:
@@ -47,7 +76,9 @@ def styled_command(cmd: str, *args: str) -> str:
     #     res.append(styled_arg(arg))
     # return " ".join(res)
 
-    return " ".join([
-        Style.GREEN(cmd),
-        *(styled_arg(f'"{arg}"' if " " in arg else arg) for arg in args),
-    ])
+    return " ".join(
+        [
+            Style.GREEN(cmd),
+            *(styled_arg(f'"{arg}"' if " " in arg else arg) for arg in args),
+        ]
+    )
